Use int instead of long for struct dir->items
2 billion files should be enough for everyone. You probably won't have
enough memory to scan such a filesystem. int is a better choice than
long, as sizeof(int) is 4 on pretty much any system where ncdu runs.

+++ src/dir.h
@@ -80,7 +80,7 @@ struct dir_output {
   /* The output code is responsible for updating these stats. Can be 0 when not
    * available. */
   int64_t size;
-  long items;
+  int items;
 };
