+++ src/util.c
@@ -283,7 +283,7 @@ struct dir *getroot(struct dir *d) {
 }
 
 
-void addparentstats(struct dir *d, int64_t size, int64_t asize, long items) {
+void addparentstats(struct dir *d, int64_t size, int64_t asize, int items) {
   while(d) {
     d->size += size;
     d->asize += asize;
